+++ include/ArduinoJson/Arduino/String.hpp
@@ -9,16 +9,7 @@
 #ifndef ARDUINO
 
 #include <string>
-
-// This class reproduces Arduino's String class
-class String : public std::string {
- public:
-  String(const char *cstr = "") : std::string(cstr) {}
-  String(const String &str) : std::string(str) {}
-  explicit String(long);
-  explicit String(int);
-  explicit String(double, unsigned char decimalPlaces = 2);
-};
+typedef std::string String;
 
 #else